Create SalesQuestion
This is the solution to a tutorial exercise completed for Computer Science course. The problem is as follows:

You have just started a sales job in a department store. Your pay consists of a base salary plus a commission. The base salary is $5,000.  Your goal is to earn $30,000 a year. Write a program that finds the minimum amount of sales you have to generate in order to make $30,000.